+++ Src/FluentAssertions/Configuration/TestFramework.cs
@@ -1,0 +1,14 @@
+namespace FluentAssertions.Configuration;
+
+/// <summary>
+/// The test frameworks supported by Fluent Assertions.
+/// </summary>
+public enum TestFramework
+{
+    XUnit2,
+    XUnit3,
+    TUnit,
+    MsTest,
+    NUnit,
+    MSpec
+}