+++ Src/FluentAssertions/ObjectAssertionsExtensions.cs
@@ -38,7 +38,7 @@ public static class ObjectAssertionsExtensions
     /// A reference to the <see cref="EquivalencyOptions{TExpectation}"/> configuration object that can be used
     /// to influence the way the object graphs are compared. You can also provide an alternative instance of the
     /// <see cref="EquivalencyOptions{TExpectation}"/> class. The global defaults are determined by the
-    /// <see cref="AssertionOptions"/> class.
+    /// <see cref="AssertionConfiguration"/> class.
     /// </param>
     /// <param name="because">
     /// A formatted phrase as is supported by <see cref="string.Format(string,object[])" /> explaining why the assertion
@@ -58,7 +58,7 @@ public static class ObjectAssertionsExtensions
         {
             var deserializedObject = CreateCloneUsingDataContractSerializer(assertions.Subject);
 
-            EquivalencyOptions<T> defaultOptions = AssertionOptions.CloneDefaults<T>()
+            EquivalencyOptions<T> defaultOptions = AssertionConfiguration.Current.Equivalency.CloneDefaults<T>()
                 .PreferringRuntimeMemberTypes().IncludingFields().IncludingProperties();
 
             ((T)deserializedObject).Should().BeEquivalentTo((T)assertions.Subject, _ => options(defaultOptions));