+++ Tests/FluentAssertions.Specs/Configuration/EquivalencyOptionsSpecs.cs
@@ -6,45 +6,43 @@ using Chill;
 using FluentAssertions.Equivalency;
 using FluentAssertions.Equivalency.Steps;
 using FluentAssertions.Execution;
-using FluentAssertions.Formatting;
+using JetBrains.Annotations;
 using Xunit;
 using Xunit.Sdk;
 
-namespace FluentAssertions.Specs;
+namespace FluentAssertions.Specs.Configuration;
 
-public class AssertionOptionsSpecs
+public class EquivalencyOptionsSpecs
 {
-    // Due to tests that call AssertionOptions
-    [CollectionDefinition("AssertionOptionsSpecs", DisableParallelization = true)]
-    public class AssertionOptionsSpecsDefinition;
-
+    [Collection("ConfigurationSpecs")]
     public abstract class Given_temporary_global_assertion_options : GivenWhenThen
     {
         protected override void Dispose(bool disposing)
         {
-            AssertionOptions.AssertEquivalencyUsing(_ => new EquivalencyOptions());
+            AssertionConfiguration.Current.Equivalency.Modify(_ => new EquivalencyOptions());
 
             base.Dispose(disposing);
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
+    [Collection("ConfigurationSpecs")]
     public class When_injecting_a_null_configurer : GivenSubject<EquivalencyOptions, Action>
     {
         public When_injecting_a_null_configurer()
         {
-            When(() => () => AssertionOptions.AssertEquivalencyUsing(defaultsConfigurer: null));
+            When(() => () => AssertionConfiguration.Current.Equivalency.Modify(configureOptions: null));
         }
 
         [Fact]
         public void It_should_throw()
         {
             Result.Should().ThrowExactly<ArgumentNullException>()
-                .WithParameterName("defaultsConfigurer");
+                .WithParameterName("configureOptions");
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
+    [Collection("ConfigurationSpecs")]
+
     public class When_concurrently_getting_equality_strategy : GivenSubject<EquivalencyOptions, Action>
     {
         public When_concurrently_getting_equality_strategy()
@@ -68,7 +66,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_modifying_global_reference_type_settings_a_previous_assertion_should_not_have_any_effect
         : Given_temporary_global_assertion_options
     {
@@ -80,7 +77,7 @@ public class AssertionOptionsSpecs
                 new MyValueType { Value = 1 }.Should().BeEquivalentTo(new MyValueType { Value = 2 });
             });
 
-            When(() => AssertionOptions.AssertEquivalencyUsing(o => o.ComparingByMembers<MyValueType>()));
+            When(() => AssertionConfiguration.Current.Equivalency.Modify(o => o.ComparingByMembers<MyValueType>()));
         }
 
         [Fact]
@@ -101,7 +98,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_modifying_global_value_type_settings_a_previous_assertion_should_not_have_any_effect
         : Given_temporary_global_assertion_options
     {
@@ -113,7 +109,7 @@ public class AssertionOptionsSpecs
                 new MyClass { Value = 1 }.Should().BeEquivalentTo(new MyClass { Value = 1 });
             });
 
-            When(() => AssertionOptions.AssertEquivalencyUsing(o => o.ComparingByValue<MyClass>()));
+            When(() => AssertionConfiguration.Current.Equivalency.Modify(o => o.ComparingByValue<MyClass>()));
         }
 
         [Fact]
@@ -138,7 +134,7 @@ public class AssertionOptionsSpecs
         {
             When(() =>
             {
-                AssertionOptions.AssertEquivalencyUsing(
+                AssertionConfiguration.Current.Equivalency.Modify(
                     options => options.ComparingByValue(typeof(Position)));
             });
         }
@@ -151,6 +147,7 @@ public class AssertionOptionsSpecs
 
         private record Position
         {
+            [UsedImplicitly]
             private readonly int value;
 
             public Position(int value)
@@ -160,14 +157,13 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_assertion_doubles_should_always_allow_small_deviations : Given_temporary_global_assertion_options
     {
         public When_assertion_doubles_should_always_allow_small_deviations()
         {
             When(() =>
             {
-                AssertionOptions.AssertEquivalencyUsing(options => options
+                AssertionConfiguration.Current.Equivalency.Modify(options => options
                     .Using<double>(ctx => ctx.Subject.Should().BeApproximately(ctx.Expectation, 0.01))
                     .WhenTypeIs<double>());
             });
@@ -192,14 +188,13 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_local_similar_options_are_used : Given_temporary_global_assertion_options
     {
         public When_local_similar_options_are_used()
         {
             When(() =>
             {
-                AssertionOptions.AssertEquivalencyUsing(options => options
+                AssertionConfiguration.Current.Equivalency.Modify(options => options
                     .Using<double>(ctx => ctx.Subject.Should().BeApproximately(ctx.Expectation, 0.01))
                     .WhenTypeIs<double>());
             });
@@ -244,7 +239,7 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
+    [Collection("ConfigurationSpecs")]
     public class Given_self_resetting_equivalency_plan : GivenWhenThen
     {
         protected override void Dispose(bool disposing)
@@ -255,11 +250,10 @@ public class AssertionOptionsSpecs
 
         protected static EquivalencyPlan Plan
         {
-            get { return AssertionOptions.EquivalencyPlan; }
+            get { return AssertionConfiguration.Current.Equivalency.Plan; }
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_inserting_a_step : Given_self_resetting_equivalency_plan
     {
         public When_inserting_a_step()
@@ -276,7 +270,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_inserting_a_step_before_another : Given_self_resetting_equivalency_plan
     {
         public When_inserting_a_step_before_another()
@@ -294,7 +287,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_appending_a_step : Given_self_resetting_equivalency_plan
     {
         public When_appending_a_step()
@@ -312,7 +304,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_appending_a_step_after_another : Given_self_resetting_equivalency_plan
     {
         public When_appending_a_step_after_another()
@@ -330,7 +321,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_appending_a_step_and_no_builtin_steps_are_there : Given_self_resetting_equivalency_plan
     {
         public When_appending_a_step_and_no_builtin_steps_are_there()
@@ -351,7 +341,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_removing_a_specific_step : Given_self_resetting_equivalency_plan
     {
         public When_removing_a_specific_step()
@@ -366,7 +355,6 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
     public class When_removing_a_specific_step_that_doesnt_exist : Given_self_resetting_equivalency_plan
     {
         public When_removing_a_specific_step_that_doesnt_exist()
@@ -381,42 +369,7 @@ public class AssertionOptionsSpecs
         }
     }
 
-    [Collection("AssertionOptionsSpecs")]
-    public class When_global_formatting_settings_are_modified : GivenWhenThen
-    {
-        private FormattingOptions oldSettings;
-
-        public When_global_formatting_settings_are_modified()
-        {
-            Given(() => oldSettings = AssertionOptions.FormattingOptions.Clone());
-
-            When(() =>
-            {
-                AssertionOptions.FormattingOptions.UseLineBreaks = true;
-                AssertionOptions.FormattingOptions.MaxDepth = 123;
-                AssertionOptions.FormattingOptions.MaxLines = 33;
-            });
-        }
-
-        [Fact]
-        public void Then_the_current_assertion_scope_should_use_these_settings()
-        {
-            AssertionScope.Current.FormattingOptions.UseLineBreaks.Should().BeTrue();
-            AssertionScope.Current.FormattingOptions.MaxDepth.Should().Be(123);
-            AssertionScope.Current.FormattingOptions.MaxLines.Should().Be(33);
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            AssertionOptions.FormattingOptions.MaxDepth = oldSettings.MaxDepth;
-            AssertionOptions.FormattingOptions.UseLineBreaks = oldSettings.UseLineBreaks;
-            AssertionOptions.FormattingOptions.MaxLines = oldSettings.MaxLines;
-
-            base.Dispose(disposing);
-        }
-    }
-
-    internal class MyEquivalencyStep : IEquivalencyStep
+    private class MyEquivalencyStep : IEquivalencyStep
     {
         public EquivalencyResult Handle(Comparands comparands, IEquivalencyValidationContext context,
             IValidateChildNodeEquivalency valueChildNodes)